Fixes, added fixtures for pack tests
