+++ src/cli/__tests__/args.test.ts
@@ -1,5 +1,5 @@
-import { check } from '../../cli/args';
 import { CLI } from '../../interfaces';
+import { check } from '../args';
 
 describe('check', () => {
   it('should return true for valid arguments', () => {
