+++ src/core/__fixtures__/pack/files/v1/file2
