+++ src/core/__fixtures__/pack/files/v2/file1
